requirements.txt
datasets>=2.17
transformers>=4.41
accelerate>=0.30
torch>=2.1
tokenizers>=0.15
svgpathtools>=1.6
svgelements>=1.9.6
beautifulsoup4>=4.12
numpy>=1.26
pillow>=10.2
tqdm>=4.66
cairosvg>=2.7
lpips>=0.1.4
torchmetrics>=1.4
open-clip-torch>=2.24
scikit-learn>=1.4
# Optional: pydiffvg (often needs system build toolchain)
